pyproject.toml
[project]
name = "letterboxd-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "letterboxdpy",
    "psycopg2>=2.9.10",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
bot = "letterboxd_discord_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.1",
]

[tool.ruff]
lint.extend-select = ["I"] # enable import sorting

[tool.uv.sources]
letterboxdpy = { git = "https://github.com/f0e/letterboxdpy.git", branch = "fix-lxml" }
